refactor: reorganize project structure into frontend and backend architecture
