+++ backend/package.json
@@ -1,23 +1,27 @@
 {
-  "name": "study-plan",
+  "name": "backend",
   "version": "1.0.0",
+  "description": "",
   "main": "index.js",
   "scripts": {
+    "test": "jest",
     "start": "node server.js",
-    "test": "node --test"
+    "dev": "nodemon server.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "description": "",
+  "type": "commonjs",
   "dependencies": {
-    "@google/genai": "^1.49.0",
+    "@google/genai": "^2.8.0",
     "cors": "^2.8.6",
     "dotenv": "^17.4.2",
     "express": "^5.2.1",
+    "mongoose": "^9.6.3",
     "sqlite3": "^6.0.1"
   },
-  "engines": {
-    "node": ">=20.x"
+  "devDependencies": {
+    "jest": "^30.4.2",
+    "nodemon": "^3.1.14"
   }
 }
