+++ frontend/src/pages/support-page/terms.html
