+++ frontend/src/utils/aiMock.js
